JCommander and JSON
Added JCommander capability to parser run time arguments, was my first time using it so I had a little play with the paramters and what the possibilities are.

Also added the ability to communicate with the server in JSON format instead of String, using Gson to parse Maps onto JSON.

Changed the database format from a simple array to a HashMap to have the ability to store a larger number of elements and simplifying adding/removing/getting them. Will hopefully experiment with using JSON as the actual database on the next stage